bot/requirements.txt
pyaes==1.6.1
pyrogram==2.0.106
pysocks==1.7.1
python-dotenv==1.0.1
